Update logic for transformation filenames
Initially I used straight forward approach replace(), however it is not very
efficient. Another way is to use regex - and I tried it on different branch - it
is nice and clean, but it increase the size of binary to 500K?! Even with all
optimizations it is still big. So, here I tried a different approach - map and
filter. I believe it is efficient and clean solution.